+++ src/Views/Error/404.aspx.designer.cs
@@ -1,0 +1,27 @@
+﻿//------------------------------------------------------------------------------
+// <generado automáticamente>
+//     Este código fue generado por una herramienta.
+//
+//     Los cambios en este archivo podrían causar un comportamiento incorrecto y se perderán si
+//     se vuelve a generar el código. 
+// </generado automáticamente>
+//------------------------------------------------------------------------------
+
+namespace lovis.Controllers.Error {
+    
+    
+    public partial class _404 {
+        
+        /// <summary>
+        /// Propiedad Master.
+        /// </summary>
+        /// <remarks>
+        /// Propiedad generada automáticamente.
+        /// </remarks>
+        public new lovis.Views.Shared.Master Master {
+            get {
+                return ((lovis.Views.Shared.Master)(base.Master));
+            }
+        }
+    }
+}