+++ src/Views/Login/Login.aspx.cs
@@ -16,11 +16,17 @@ namespace lovis.Views.Login
         {
             if (!IsPostBack)
             {
-                /// Test acc
-                Controllers.Users.Users.uLTest.Clear();
-                new Controllers.Users.Users("test@test.com", "test", "Emiliano", "Montesdeoca").New();
-                //new Controllers.Users.Users("test2@test.com", "test", "Luis", "Sanfiel").New();
 
+                /// Test accs
+                if (Controllers.Users.Users.uLTest.Exists(x => Security.CryptoUtils.DecodeUsername(x.Username) == "test@test.com"))
+                {
+
+                }
+                else
+                {
+                    new Controllers.Users.Users("test@test.com", "test", "Emiliano", "Montesdeoca").New();
+
+                }
 
             }
             else
@@ -28,7 +34,6 @@ namespace lovis.Views.Login
                 Session["User"] = null;
 
             }
-
         }
 
         protected void login_loginbtn_Click(object sender, EventArgs e)
@@ -36,7 +41,17 @@ namespace lovis.Views.Login
             Controllers.Users.Users u = new Controllers.Users.Users().isUser(login_username.Text, login_password.Text);
             if (u != null)
             {
-                Session["User"] = new Controllers.Users.Users().DecodeUser(u);
+                Controllers.Users.Users loginUser = new Controllers.Users.Users();
+
+                loginUser.Id = u.Id;
+                loginUser.Username = Security.CryptoUtils.DecodeUsername(u.Username);
+                loginUser.Name = Security.CryptoUtils.DecodeUserString(u.Name, u);
+                loginUser.Surname = Security.CryptoUtils.DecodeUserString(u.Surname, u);
+                loginUser.DateRegister = u.DateRegister;
+                loginUser.PasswordHash = u.PasswordHash;
+                loginUser.RandomKey = u.RandomKey;
+
+                Session["User"] = loginUser;
                 Response.Redirect("~/Views/Dashboard/Dashboard.aspx");
             }