+++ src/Views/Proyects/Proyects.aspx.designer.cs
@@ -13,6 +13,33 @@ namespace lovis.Views.Proyects {
     public partial class Proyects {
         
         /// <summary>
+        /// Control pagetitle.
+        /// </summary>
+        /// <remarks>
+        /// Campo generado automáticamente.
+        /// Para modificarlo, mueva la declaración del campo del archivo del diseñador al archivo de código subyacente.
+        /// </remarks>
+        protected global::System.Web.UI.WebControls.Literal pagetitle;
+        
+        /// <summary>
+        /// Control proyect_summary.
+        /// </summary>
+        /// <remarks>
+        /// Campo generado automáticamente.
+        /// Para modificarlo, mueva la declaración del campo del archivo del diseñador al archivo de código subyacente.
+        /// </remarks>
+        protected global::System.Web.UI.HtmlControls.HtmlGenericControl proyect_summary;
+        
+        /// <summary>
+        /// Control proyect_people.
+        /// </summary>
+        /// <remarks>
+        /// Campo generado automáticamente.
+        /// Para modificarlo, mueva la declaración del campo del archivo del diseñador al archivo de código subyacente.
+        /// </remarks>
+        protected global::System.Web.UI.HtmlControls.HtmlGenericControl proyect_people;
+        
+        /// <summary>
         /// Propiedad Master.
         /// </summary>
         /// <remarks>